RecordCellを他のファイルからも参照できるようにする
